% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indels.R
\name{compute_indels}
\alias{compute_indels}
\title{Compute Indels}
\usage{
compute_indels(seq_data, indel_thres = 0.1)
}
\arguments{
\item{seq_data}{A data frame of sequencing results.}

\item{indel_thres}{Numeric threshold for identifying indels.}
}
\value{
A data frame with indel counts.
}
\description{
Calculates indel frequencies from sequencing data.
}
