% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_gRNA_specificity.R
\name{evaluate_gRNA_specificity}
\alias{evaluate_gRNA_specificity}
\title{Evaluate gRNA Specificity}
\usage{
evaluate_gRNA_specificity(gRNA_seq, off_target_seqs)
}
\arguments{
\item{gRNA_seq}{A character string representing the guide RNA sequence.}

\item{off_target_seqs}{A character vector of potential off-target sequences.}
}
\value{
A data frame with off-target sequences and their similarity scores.
}
\description{
This function calculates the specificity score of a guide RNA (gRNA) sequence
by comparing it with a list of potential off-target sequences. The specificity
score represents the similarity between the gRNA and each off-target sequence,
helping to identify sequences that may lead to unintended edits.
Higher specificity scores indicate a greater likelihood of off-target binding.
}
