% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_off_target_map.R
\name{plot_off_target_map}
\alias{plot_off_target_map}
\title{Plot Off-Target Effects}
\usage{
plot_off_target_map(off_target_data)
}
\arguments{
\item{off_target_data}{Data frame with information on off-target sites.}
}
\value{
A plotly object for interactive exploration.
}
\description{
Creates an interactive map of off-target effects across the genome.
}
