% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gRNA_base_composition.R
\name{visualize_gRNA_base_composition}
\alias{visualize_gRNA_base_composition}
\title{Visualize gRNA Base Composition}
\usage{
visualize_gRNA_base_composition(gRNA_seqs)
}
\arguments{
\item{gRNA_seqs}{A character vector of gRNA sequences.}
}
\value{
A ggplot object showing base composition.
}
\description{
This function generates a bar plot displaying the base composition
(A, T, G, C) of a set of guide RNA (gRNA) sequences. By summarizing
the frequency of each base, this plot provides insights into the nucleotide
makeup of the gRNA sequences, which can impact binding stability and efficiency.
The output is a ggplot object, making it easy to customize or integrate into
reports.
}
