% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gRNA_efficiency.R
\name{visualize_gRNA_efficiency}
\alias{visualize_gRNA_efficiency}
\title{Visualize gRNA Efficiency}
\usage{
visualize_gRNA_efficiency(efficiency_data)
}
\arguments{
\item{efficiency_data}{A data frame with columns for gRNA and efficiency scores.}
}
\value{
A ggplot object displaying the heatmap.
}
\description{
This function creates a heatmap to visually compare the efficiency
of multiple guide RNA (gRNA) sequences. Each gRNA is represented
with an efficiency score, and the heatmap provides a quick visual assessment
of high and low efficiency sequences, aiding in the selection
of optimal gRNAs for experiments. The output is a ggplot object,
making it easy to customize or integrate into reports.
}
